Renamed a file
refs: #3199
