+++ translations/dnc_core.js
